Disable Commander X16 BASIC mode by default
The BASIC extension for output mode is select by the file headers, not the
command line, and selecting X16 mode by default breaks compatibility with
previous versions that would use TFC 3 BASIC (not that anyone would use
that, but it was what I had back when I wrote this program 25 years ago...)

Write a special "tokx16" header when converting to text, and use this to
override autodetection. Allow -x in input mode to decode X16 BASIC with a
standard tok64 header.

Clean up the help screen to make this clearer, and document the -x option
in the README.