Update documentation
Update the manual page to match recent updates.

In commit fcd0ef34f692451 from 1999 (from before I used Git) I
committed a lot of changes that had been around in my working tree for
a year or so without me remembering what I did. Apparently, part of
that was that I intended to add support for a lot of BASIC extensions,
but I never did. I did add them to the documentation, and promptly
forgot about them.

Remove them--do not claim support for something that is not
implemented.